refactor(enrichment): extract SSRF guards into shared url-guard module
