+++ apps/core/src/modules/enrichment/providers/open-graph/safe-fetch.ts
@@ -1,7 +1,8 @@
-import { lookup } from 'node:dns/promises'
-import { isIP } from 'node:net'
-
 import { isDev } from '~/global/env.global'
+
+import { assertHostnameSafe, parseAndValidateUrl } from './url-guard'
+
+export { isPrivateIp, UnsafeUrlError } from './url-guard'
 
 const MAX_REDIRECTS = 5
 // Pose as a current Chrome on macOS. Self-identifying "bot" UAs (the previous
@@ -12,16 +13,6 @@ const DEFAULT_UA =
   'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/131.0.0.0 Safari/537.36'
 const SEC_CH_UA =
   '"Google Chrome";v="131", "Chromium";v="131", "Not_A Brand";v="24"'
-
-const BLOCKED_HOSTNAMES = new Set([
-  'localhost',
-  'localhost.localdomain',
-  'broadcasthost',
-  'ip6-localhost',
-  'ip6-loopback',
-])
-
-const BLOCKED_HOSTNAME_SUFFIXES = ['.localhost', '.local', '.internal']
 
 export interface SafeFetchOptions {
   timeoutMs: number
@@ -35,13 +26,6 @@ export interface SafeFetchResult {
   contentType: string
   body: string
   truncated: boolean
-}
-
-export class UnsafeUrlError extends Error {
-  constructor(message: string) {
-    super(message)
-    this.name = 'UnsafeUrlError'
-  }
 }
 
 export class FetchSizeExceededError extends Error {
@@ -177,64 +161,6 @@ function buildRequestHeaders(
   }
 }
 
-function parseAndValidateUrl(raw: string): URL {
-  let url: URL
-  try {
-    url = new URL(raw)
-  } catch {
-    throw new UnsafeUrlError(`Invalid URL: ${raw}`)
-  }
-  if (url.protocol !== 'http:' && url.protocol !== 'https:') {
-    throw new UnsafeUrlError(`Disallowed protocol: ${url.protocol}`)
-  }
-  const host = url.hostname.toLowerCase()
-  if (!host) throw new UnsafeUrlError('Empty hostname')
-  // Dev escape hatch: skip private-network guards so local proxies that
-  // route public domains through fake-IP / RFC2544 ranges (e.g. Surge's
-  // 198.18/15 enhanced-mode pool) can still resolve through this fetcher.
-  if (isDev) return url
-  if (BLOCKED_HOSTNAMES.has(host)) {
-    throw new UnsafeUrlError(`Blocked hostname: ${host}`)
-  }
-  if (BLOCKED_HOSTNAME_SUFFIXES.some((s) => host.endsWith(s))) {
-    throw new UnsafeUrlError(`Blocked hostname suffix: ${host}`)
-  }
-  const ipKind = isIP(stripIpv6Brackets(host))
-  if (ipKind !== 0 && isPrivateIp(stripIpv6Brackets(host), ipKind)) {
-    throw new UnsafeUrlError(`Private IP literal: ${host}`)
-  }
-  return url
-}
-
-async function assertHostnameSafe(hostname: string): Promise<void> {
-  // Skip DNS lookup for IP literals (already validated in parse step).
-  if (isIP(stripIpv6Brackets(hostname)) !== 0) return
-
-  let addrs: { address: string; family: number }[]
-  try {
-    addrs = await lookup(hostname, { all: true })
-  } catch (error) {
-    throw new UnsafeUrlError(
-      `DNS lookup failed for ${hostname}: ${(error as Error).message}`,
-    )
-  }
-  if (addrs.length === 0) {
-    throw new UnsafeUrlError(`No DNS records for ${hostname}`)
-  }
-  for (const a of addrs) {
-    if (isPrivateIp(a.address, a.family === 4 ? 4 : 6)) {
-      throw new UnsafeUrlError(
-        `Hostname ${hostname} resolves to private/internal IP ${a.address}`,
-      )
-    }
-  }
-}
-
-function stripIpv6Brackets(host: string): string {
-  if (host.startsWith('[') && host.endsWith(']')) return host.slice(1, -1)
-  return host
-}
-
 function isRedirect(status: number): boolean {
   return (
     status === 301 ||
@@ -296,54 +222,4 @@ async function readWithLimit(
     }
   }
   return { body: chunks.join(''), truncated }
-}
-
-/**
- * IPv4: blocks loopback (127/8), private (10/8, 172.16/12, 192.168/16),
- * link-local (169.254/16), CGNAT (100.64/10), broadcast (255.255.255.255),
- * "this network" (0/8) and multicast (224/4).
- *
- * IPv6: blocks loopback (::1), unspecified (::), link-local (fe80::/10),
- * unique-local (fc00::/7), IPv4-mapped private equivalents, and multicast.
- */
-export function isPrivateIp(addr: string, family: number): boolean {
-  if (family === 4) return isPrivateIpv4(addr)
-  if (family === 6) return isPrivateIpv6(addr)
-  return true
-}
-
-function isPrivateIpv4(addr: string): boolean {
-  const parts = addr.split('.').map((p) => Number(p))
-  if (
-    parts.length !== 4 ||
-    parts.some((p) => !Number.isInteger(p) || p < 0 || p > 255)
-  ) {
-    return true
-  }
-  const [a, b] = parts
-  if (a === 0) return true
-  if (a === 10) return true
-  if (a === 127) return true
-  if (a === 169 && b === 254) return true
-  if (a === 172 && b >= 16 && b <= 31) return true
-  if (a === 192 && b === 168) return true
-  if (a === 192 && b === 0) return true // 192.0.0.0/24, 192.0.2.0/24 doc range
-  if (a === 198 && (b === 18 || b === 19)) return true // benchmark
-  if (a === 100 && b >= 64 && b <= 127) return true // CGNAT
-  if (a >= 224) return true // multicast + reserved
-  return false
-}
-
-function isPrivateIpv6(addr: string): boolean {
-  const lower = addr.toLowerCase()
-  if (lower === '::' || lower === '::1') return true
-  // IPv4-mapped — re-check as v4.
-  const v4mapped = /^:{2}f{4}:((?:\d+\.){3}\d+)$/i.exec(addr)
-  if (v4mapped) return isPrivateIpv4(v4mapped[1])
-  // fe80::/10 link-local, fc00::/7 unique-local, ff00::/8 multicast.
-  if (lower.startsWith('fe8') || lower.startsWith('fe9')) return true
-  if (lower.startsWith('fea') || lower.startsWith('feb')) return true
-  if (lower.startsWith('fc') || lower.startsWith('fd')) return true
-  if (lower.startsWith('ff')) return true
-  return false
 }
